+++ spec/classes/config_spec.rb
@@ -2,7 +2,13 @@ require 'spec_helper'
 
 describe 'ejabberd::config' do
 
-    it { should contain_file('/etc/ejabberd/ejabberd.cfg') }
-    it { should contain_exec('reload-config') }
+    context 'with default config' do
+        let(:params) { {:config => "template('ejabberd/ejabberd.cfg.erb')" } }
+
+        it { should contain_file('/etc/ejabberd/ejabberd.cfg') \
+            .with_content("template('ejabberd/ejabberd.cfg.erb')")
+        }
+        it { should contain_exec('reload-config') }
+    end
 
 end