+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.2 2014-xx-xx
 - Calls ejabberdctl load_config to reload ACLs when config changes. This means that ejabberd is no longer restarted when config changes.
+- Allow ejabberd config file to be specified
 - Some specs added
 
 ## 0.0.1 2014-01-10